// src/interfaces/supabase/home.ts
export interface IHomeInformation {
    id: number;
    language_uuid: string;
    main_title: string;
    created_at: string;
    history: string;
    our_team: string;
    main_subtitles: string;
    about_us_title: string;
    about_us_description: string;
    contact_card_title: string;
    contact_card_subtitle: string;
}